+++ .eslintrc.js
@@ -14,8 +14,6 @@ module.exports = {
     'no-unused-vars': ['off'],
     'react-native/no-unused-styles': 2,
     'react-native/split-platform-components': 2,
-    'react-native/no-inline-styles': 2,
-    'react-native/no-color-literals': 2,
   },
   parserOptions: {
     ecmaVersion: 6,