+++ index.d.ts
@@ -1,1 +1,1 @@
-declare module 'color-sort'
+declare module 'color-sort';