+++ lib/components/Item.d.ts
@@ -1,2 +1,2 @@
-import { Item as ItemProps } from '../interfaces';
-export declare const Item: import("styled-components").StyledComponent<typeof import("react-native").TouchableOpacity, import("styled-components").DefaultTheme, ItemProps, never>;
+import { ItemProps } from '../interfaces';
+export declare const Item: ({ children, style, color, itemSize, shadow, ...props }: ItemProps) => JSX.Element;