+++ lib/components/Marker.d.ts
@@ -1,2 +1,2 @@
-import { Marker as MarkerProps } from '../interfaces';
-export declare const Marker: ({ animate, color, size, markerStyle, type, }: MarkerProps) => JSX.Element;
+import { MarkerProps } from '../interfaces';
+export declare const Marker: ({ animate, color, size, display, type, }: MarkerProps) => JSX.Element;