+++ lib/helpers/mixin.d.ts
@@ -1,1 +1,0 @@
-export declare const mixin = "\n  position: absolute;\n  top: 0;\n  left: 0;\n  bottom: 0;\n  right: 0;\n";