+++ lib/interfaces/Gradient.d.ts
@@ -1,4 +1,6 @@
-import { LinearGradientProps } from 'expo';
-export interface Gradient extends LinearGradientProps {
+import { StyleProp, ViewStyle } from 'react-native';
+export interface Gradient {
     readonly size: number;
+    readonly colors: string[];
+    readonly style: StyleProp<ViewStyle>;
 }