+++ lib/types/MarkerDisplay.d.ts
@@ -1,0 +1,1 @@
+export declare type MarkerDisplay = 'adjust' | 'contrast' | string;