+++ src/components/index.tsx
@@ -1,4 +1,3 @@
 export { Gradient } from './Gradient'
 export { Item } from './Item'
 export { Marker } from './Marker'
-export { Props } from './Props'