+++ src/helpers/index.ts
@@ -1,3 +1,2 @@
 export { checkColor, darker, lighter } from './colors'
 export { countSize } from './size'
-export { mixin } from './mixin'