+++ src/helpers/mixin.tsx
@@ -1,7 +1,0 @@
-export const mixin = `
-  position: absolute;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  right: 0;
-`