+++ src/interfaces/Gradient.ts
@@ -1,0 +1,7 @@
+import { StyleProp, ViewStyle } from 'react-native'
+
+export interface Gradient {
+  readonly size: number
+  readonly colors: string[]
+  readonly style: StyleProp<ViewStyle>
+}