+++ src/interfaces/Gradient.tsx
@@ -1,5 +1,0 @@
-import { LinearGradientProps } from 'expo'
-
-export interface Gradient extends LinearGradientProps {
-  readonly size: number
-}