+++ src/interfaces/index.ts
@@ -1,0 +1,4 @@
+export { Gradient as GradientProps } from './Gradient'
+export { Item as ItemProps } from './Item'
+export { Marker as MarkerProps } from './Marker'
+export { Props } from './Props'