+++ src/types/index.ts
@@ -1,0 +1,4 @@
+export { Animate } from './Animate'
+export { MarkerType } from './MarkerType'
+export { MarkerDisplay } from './MarkerDisplay'
+export { Children } from './Children'