+++ src/types/index.tsx
@@ -1,3 +1,0 @@
-export { Animate } from './Animate'
-export { Marker } from './Marker'
-export { MarkerStyle } from './MarkerStyle'